add jdbc plugin support
Add support for using a jdbc plugin instead of the sqlserver
driver that is packaged with the plugin. For example, somebody
could use the jtds driver if they want.

To support this, added config properties for the jdbc plugin
name and connection string. Copied relevant classes and methods
from db-plugins into DBUtils, DriverCleanup, and JDBCDriverShim
to support custom jdbc.

Also fixed a bug where the CDC timestamp was not being included
in the output records.